refactor: mark new helpers as `unstable_`
